+++ apps/frontend/src/app/(blog)/blog/[slug]/page.tsx
@@ -1,5 +1,13 @@
+import '@/styles/blog.scss';
+import { Metadata } from 'next';
+import Image from 'next/image';
+import { notFound } from 'next/navigation';
+import {
+  Avatar,
+  AvatarFallback,
+  AvatarImage,
+} from '../../../../components/old/avatar';
 import { SyntaxHighlight } from '../../../../components/syntax-highlight';
-import { Avatar, AvatarFallback, AvatarImage } from '../../../../components/ui/avatar';
 import { Heading } from '../../../../components/ui/heading';
 import { getCmsAssetUrl } from '../../../../lib/cms-asset';
 import { renderMarkdown } from '../../../../lib/markdown';
@@ -7,10 +15,6 @@ import { imageFallback } from '../../../../lib/post-image-fallback';
 import { nameLetters } from '../../../../lib/utils';
 import { getSeoData } from '../../../../services/cms-api';
 import { getPost } from '../../../../services/cms-api/apis/blog';
-import '@/styles/blog.scss';
-import { Metadata } from 'next';
-import Image from 'next/image';
-import { notFound } from 'next/navigation';
 
 type Props = {
   params: {
