+++ apps/frontend/src/app/(landing)/@contact/send-button.tsx
@@ -1,7 +1,7 @@
-import { Button } from '../../../components/ui/button';
-import { cn } from '../../../lib/utils';
 import { ComponentProps } from 'react';
 import { FaPaperPlane, FaSpinner } from 'react-icons/fa6';
+import { Button } from '../../../components/old/button';
+import { cn } from '../../../lib/utils';
 
 type Props = ComponentProps<typeof Button> & {
   isLoading?: boolean;
