+++ apps/frontend/src/app/(landing)/@experience/components/experience-card/unemployement-card.tsx
@@ -1,5 +1,5 @@
-import { Button } from '../../../../../components/ui/button';
 import Link from 'next/link';
+import { Button } from '../../../../../components/old/button';
 import { RocketAnimation } from '../rocket-animation/rocket-animation';
 
 export default function UnemployementCard() {
