+++ apps/frontend/src/app/sections/navbar/drawer.tsx
@@ -1,16 +1,16 @@
 'use client';
 
-import { Button } from '../../../components/ui/button';
+import { Button } from '../../../components/old/button';
 
+import Link from 'next/link';
+import { FaBars, FaXmark } from 'react-icons/fa6';
 import {
   Sheet,
   SheetClose,
   SheetContent,
   SheetTitle,
   SheetTrigger,
-} from '../../../components/ui/sheet';
-import Link from 'next/link';
-import { FaBars, FaXmark } from 'react-icons/fa6';
+} from '../../../components/old/sheet';
 
 export function NavbarDrawer() {
   return (
