+++ apps/frontend/src/app/sections/navbar/navbar.tsx
@@ -1,6 +1,6 @@
-import { Button } from '../../../components/ui/button';
 import Image from 'next/image';
 import Link from 'next/link';
+import { Button } from '../../../components/old/button';
 import { NavbarDrawer } from './drawer';
 
 export function Navbar() {
