+++ apps/frontend/src/components/old/avatar.tsx
