+++ apps/frontend/src/components/old/carousel.tsx
